refactoring: code refactoring
This code is being refactored, in order to fix race conditions in tests
and make code base more tests-friendly.
