+++ internal/output/helpers.go
@@ -1,0 +1,8 @@
+package output
+
+import "os"
+
+func IsFileExist(path string) bool {
+	_, err := os.Stat(path)
+	return !os.IsNotExist(err)
+}
