+++ internal/output/imports.go
@@ -1,0 +1,6 @@
+package output
+
+// we do not need support for aliases in import for now.
+var importsMap = map[string]string{
+	"testify": "github.com/stretchr/testify/assert",
+}
