+++ internal/render/render.go
@@ -1,12 +1,10 @@
 package render
 
-//go:generate esc -o bindata/esc.go -pkg=bindata templates
 import (
 	"fmt"
 	"io"
 	"io/ioutil"
 	"path"
-	"strings"
 	"text/template"
 
 	"github.com/cweill/gotests/internal/models"
@@ -14,36 +12,35 @@ import (
 	"github.com/cweill/gotests/templates"
 )
 
-const (
-	name  = "name"
-	nFile = 7
-)
-
-var tmpls *template.Template
-
-func init() {
-	Reset()
+type Render struct {
+	tmpls *template.Template
 }
 
-func Reset() {
-	initEmptyTmpls()
+func New() *Render {
+	r := Render{
+		tmpls: template.New("render").Funcs(map[string]interface{}{
+			"Field":    fieldName,
+			"Receiver": receiverName,
+			"Param":    parameterName,
+			"Want":     wantName,
+			"Got":      gotName,
+		}),
+	}
+
+	// default templates first
 	for _, name := range bindata.AssetNames() {
-		tmpls = template.Must(tmpls.Parse(bindata.FSMustString(false, name)))
+		// if strings.Contains(bindata.FSMustString(false, name), "function") {
+		// 	fmt.Printf("render.New - bindata.[%s]\n", name)
+		// 	// fmt.Printf("> %s\n", bindata.FSMustString(false, name))
+		// }
+		r.tmpls = template.Must(r.tmpls.Parse(bindata.FSMustString(false, name)))
 	}
-}
 
-// LoadFromData allows to load from a data slice
-func LoadFromData(templateData [][]byte) {
-	initEmptyTmpls()
-	for _, d := range templateData {
-		tmpls = template.Must(tmpls.Parse(string(d)))
-	}
+	return &r
 }
 
 // LoadCustomTemplates allows to load in custom templates from a specified path.
-func LoadCustomTemplates(dir string) error {
-	initEmptyTmpls()
-
+func (r *Render) LoadCustomTemplates(dir string) error {
 	files, err := ioutil.ReadDir(dir)
 	if err != nil {
 		return fmt.Errorf("ioutil.ReadDir: %v", err)
@@ -53,15 +50,19 @@ func LoadCustomTemplates(dir string) error {
 	for _, f := range files {
 		templateFiles = append(templateFiles, path.Join(dir, f.Name()))
 	}
-	tmpls, err = tmpls.ParseFiles(templateFiles...)
+
+	// fmt.Printf("template(s)- %#v\n", templateFiles)
+
+	r.tmpls, err = r.tmpls.ParseFiles(templateFiles...)
 	if err != nil {
 		return fmt.Errorf("tmpls.ParseFiles: %v", err)
 	}
+
 	return nil
 }
 
 // LoadCustomTemplatesName allows to load in custom templates of a specified name from the templates directory.
-func LoadCustomTemplatesName(name string) error {
+func (r *Render) LoadCustomTemplatesName(name string) error {
 	f, err := templates.Dir(false, "/").Open(name)
 	if err != nil {
 		return fmt.Errorf("templates.Open: %v", err)
@@ -78,97 +79,39 @@ func LoadCustomTemplatesName(name string) error {
 			return fmt.Errorf("templates.FSString: %v", err)
 		}
 
-		tmpls, err = tmpls.Parse(text)
+		tmpls, err = r.tmpls.Parse(text)
 		if err != nil {
 			return fmt.Errorf("tmpls.Parse: %v", err)
 		}
+		r.tmpls = tmpls
 	}
 
 	return nil
 }
 
-func initEmptyTmpls() {
-	tmpls = template.New("render").Funcs(map[string]interface{}{
-		"Field":    fieldName,
-		"Receiver": receiverName,
-		"Param":    parameterName,
-		"Want":     wantName,
-		"Got":      gotName,
-	})
+func (r *Render) LoadFromData(templateData [][]byte) {
+	for _, d := range templateData {
+		r.tmpls = template.Must(r.tmpls.Parse(string(d)))
+	}
 }
 
-func fieldName(f *models.Field) string {
-	var n string
-	if f.IsNamed() {
-		n = f.Name
-	} else {
-		n = f.Type.String()
-	}
-	return n
-}
-
-func receiverName(f *models.Receiver) string {
-	var n string
-	if f.IsNamed() {
-		n = f.Name
-	} else {
-		n = f.ShortName()
-	}
-	if n == "name" {
-		// Avoid conflict with test struct's "name" field.
-		n = "n"
-	} else if n == "t" {
-		// Avoid conflict with test argument.
-		// "tr" is short for t receiver.
-		n = "tr"
-	}
-	return n
-}
-
-func parameterName(f *models.Field) string {
-	var n string
-	if f.IsNamed() {
-		n = f.Name
-	} else {
-		n = fmt.Sprintf("in%v", f.Index)
-	}
-	return n
-}
-
-func wantName(f *models.Field) string {
-	var n string
-	if f.IsNamed() {
-		n = "want" + strings.Title(f.Name)
-	} else if f.Index == 0 {
-		n = "want"
-	} else {
-		n = fmt.Sprintf("want%v", f.Index)
-	}
-	return n
-}
-
-func gotName(f *models.Field) string {
-	var n string
-	if f.IsNamed() {
-		n = "got" + strings.Title(f.Name)
-	} else if f.Index == 0 {
-		n = "got"
-	} else {
-		n = fmt.Sprintf("got%v", f.Index)
-	}
-	return n
-}
-
-func Header(w io.Writer, h *models.Header) error {
-	if err := tmpls.ExecuteTemplate(w, "header", h); err != nil {
+func (r *Render) Header(w io.Writer, h *models.Header) error {
+	if err := r.tmpls.ExecuteTemplate(w, "header", h); err != nil {
 		return err
 	}
 	_, err := w.Write(h.Code)
 	return err
 }
 
-func TestFunction(w io.Writer, f *models.Function, printInputs, subtests, named, parallel bool, templateParams map[string]interface{}) error {
-	return tmpls.ExecuteTemplate(w, "function", struct {
+func (r *Render) TestFunction(
+	w io.Writer,
+	f *models.Function,
+	printInputs bool,
+	subtests bool,
+	named bool,
+	parallel bool,
+	params map[string]interface{}) error {
+	return r.tmpls.ExecuteTemplate(w, "function", struct {
 		*models.Function
 		PrintInputs    bool
 		Subtests       bool
@@ -181,6 +124,6 @@ func TestFunction(w io.Writer, f *models.Function, printInputs, subtests, named,
 		Subtests:       subtests,
 		Parallel:       parallel,
 		Named:          named,
-		TemplateParams: templateParams,
+		TemplateParams: params,
 	})
 }
